pyproject.toml
[project]
name = "noetl"
version = "1.0.0"
description = "A framework to build and run data pipelines and workflows."
authors = [
    { name = "Kadyapam", email = "182583029+kadyapam@users.noreply.github.com" }
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
keywords = ["etl", "ml", "ops", "cleansing", "data", "pipeline", "workflow", "automation"]
dependencies = [
    "fastapi>=0.115.6",
    "pydantic>=2.11.4",
    "aiofiles==24.1.0",
    "psycopg[binary,pool]>=3.2.7",
    "connectorx>=0.4.3",
    "greenlet>=3.2.1",
    "uvicorn>=0.34.0",
    "requests>=2.32.3",
    "httpx>=0.28.1",
    "google-auth>=2.27.0",
    "python-multipart==0.0.20",
    "PyYAML>=6.0.1",
    "Jinja2>=3.1.6",
    "pycryptodome>=3.21",
    "urllib3>=2.3",
    "Authlib>=1.5.1",
    "typer>=0.15.3",
    "click>=8.1.0,<8.2.1",
    "psutil>=7.0.0",
    "memray>=1.17.2",
    "deepdiff>=8.4.2",
    "lark>=1.2.2",
    "duckdb>=1.3.0",
    "duckdb-engine>=0.17.0",
    "snowflake-connector-python>=4.0.0",
    "polars[pyarrow]>=1.30.0",
    "networkx>=3.5",
    "pydot>=4.0.1",
    "fsspec>=2025.5.1",
    "gcsfs>=2025.5.1",
    "boto3>=1.38.45",
    "azure-identity>=1.23.0",
    "azure-keyvault-secrets>=4.8.0",
    "ortools>=9.10"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.14.0"
]
publish = [
    "build>=1.2.2.post1",
    "twine>=6.1.0",
]
# Jupyter/notebook environment and data-viz/science stack
notebook = [
    "jupyterlab>=4.4.3",
    "jupysql>=0.11.1",
    "pandas>=2.2.3",
    "matplotlib>=3.10.3"
]

[project.scripts]
noetl = "noetl.main:main"

[project.urls]
Homepage = "https://noetl.io"
Repository = "https://github.com/noetl/noetl"
Issues = "https://github.com/noetl/noetl/issues"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["noetl*"]

[tool.setuptools.package-data]
"noetl" = ["ui/**/*", "database/ddl/**/*"]

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "twine>=6.1.0",
]
